@@ -1,6 +1,5 @@
 {
-  "name": "ui",
-  "description": "Returns a UI to interact with samples",
+  "name": "noop-local-inspector",
   "version": "0.1.0",
   "private": true,
   "scripts": {
@@ -9,28 +8,35 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.2",
-    "@fortawesome/free-brands-svg-icons": "^5.2.0",
-    "@fortawesome/free-regular-svg-icons": "^5.2.0",
-    "@fortawesome/free-solid-svg-icons": "^5.2.0",
-    "@fortawesome/vue-fontawesome": "^0.1.1",
-    "@rearc/noop-log": "^1.0.2",
-    "@vue/cli-plugin-babel": "^3.0.0-rc.11",
-    "@vue/cli-plugin-eslint": "^3.0.0-rc.11",
-    "@vue/cli-service": "^4.5.3",
-    "@vue/eslint-config-standard": "^3.0.0-rc.11",
-    "axios": "^0.18.0",
-    "bootstrap-vue": "^2.0.0-rc.11",
+    "@fortawesome/fontawesome-svg-core": "^1.2.30",
+    "@fortawesome/free-solid-svg-icons": "^5.14.0",
+    "@fortawesome/vue-fontawesome": "^0.1.10",
+    "@vue/cli-plugin-babel": "~4.5.0",
+    "@vue/cli-plugin-eslint": "~4.5.0",
+    "@vue/cli-plugin-router": "~4.5.0",
+    "@vue/cli-plugin-vuex": "~4.5.0",
+    "@vue/cli-service": "~4.5.0",
+    "@vue/eslint-config-standard": "^5.1.2",
+    "axios": "^0.20.0",
+    "babel-eslint": "^10.1.0",
+    "bootstrap-vue": "^2.16.0",
     "content-type": "^1.0.4",
+    "core-js": "^3.6.5",
+    "eslint": "^6.7.2",
+    "eslint-plugin-import": "^2.20.2",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-promise": "^4.2.1",
+    "eslint-plugin-standard": "^4.0.0",
+    "eslint-plugin-vue": "^6.2.2",
+    "moment": "^2.27.0",
     "prettysize": "^2.0.0",
-    "query-string": "^6.5.0",
-    "vue": "^2.5.17",
+    "vue": "^2.6.11",
     "vue-bootstrap-typeahead": "^0.2.6",
-    "vue-jquery": "^1.0.6",
-    "vue-router": "^3.0.2",
-    "vue-template-compiler": "^2.5.17",
-    "vuex": "^3.1.0",
-    "xterm": "^3.12.2"
+    "vue-router": "^3.2.0",
+    "vuex": "^3.4.0",
+    "xterm": "^4.8.1",
+    "xterm-addon-fit": "^0.4.0",
+    "vue-template-compiler": "^2.6.11"
   },
   "eslintConfig": {
     "root": true,
@@ -41,19 +47,14 @@
       "plugin:vue/essential",
       "@vue/standard"
     ],
-    "rules": {},
     "parserOptions": {
       "parser": "babel-eslint"
-    }
-  },
-  "postcss": {
-    "plugins": {
-      "autoprefixer": {}
-    }
+    },
+    "rules": {}
   },
   "browserslist": [
     "> 1%",
     "last 2 versions",
-    "not ie <= 8"
+    "not dead"
   ]
 }
