@@ -2,11 +2,11 @@
   <div id="app">
     <div id="top" class="container-fluid">
       <div id="header" class="row">
-        <div id="logo" class="col-auto text-left" @click="$router.push('/')">
+        <div id="logo" class="col-auto text-left" @click="$route.path === '/' ? null : $router.push('/')">
           <img alt="Rearc logo" src="./assets/logo_darkbackground.png" class="logo">
         </div>
         <div id="title" class="col text-left">
-          noop local inspector
+          Noop Local Inspector
         </div>
         <div class="coll text-right">
           <ChangesIndicator />
@@ -15,14 +15,14 @@
       <div id="spacer"></div>
     </div>
     <b-container fluid id="main">
-      <router-view/>
+      <router-view />
     </b-container>
     <div class="background"></div>
   </div>
 </template>
 
 <script>
-import ChangesIndicator from '@/components/ChangesIndicator.vue'
+import ChangesIndicator from './components/ChangesIndicator.vue'
 
 export default {
   components: {
