@@ -1,4 +1,4 @@
-import { baseUrl } from '@/api'
+import { baseUrl } from '../../api'
 
 const state = { connected: false }
 
@@ -30,7 +30,7 @@ const actions = {
       } catch (err) {
         console.error('WS msg parse error', err)
       }
-      //do something with payload
+      // do something with payload
       console.log(payload)
     }
   }
