@@ -1,4 +1,4 @@
-import { request } from '@/api'
+import { request } from '../../api'
 
 const state = { data: [] }
 
@@ -30,16 +30,16 @@ const actions = {
       })
   },
   putVariable ({ commit }, req) {
-    request.put(`/components/${req.id}/variables/${req.key}`, {value: req.value})
+    request.put(`/components/${req.id}/variables/${req.key}`, { value: req.value })
       .then(res => {
         actions.byId({ commit }, req.id)
-    })
+      })
   },
   removeVariable ({ commit }, req) {
     request.delete(`/components/${req.id}/variables/${req.key}`)
       .then(res => {
         actions.byId({ commit }, req.id)
-    })
+      })
   }
 }
 
