@@ -1,4 +1,4 @@
-import { baseUrl } from '@/api'
+import { baseUrl } from '../../api'
 
 const state = { data: [], connected: false }
 
