@@ -240,7 +240,7 @@ class Inspector {
         return {
           id: container.resource.name,
           type: container.resource.type,
-          params: container.resource.params,
+          settings: container.resource.settings,
           container: {
             image: container.getImage(),
             name: container.name
@@ -264,7 +264,7 @@ class Inspector {
       res.json({
         id: container.resource.name,
         type: container.resource.type,
-        parameters: container.resource.params,
+        settings: container.resource.settings,
         components,
         container: {
           image: container.getImage(),
@@ -348,14 +348,6 @@ class Inspector {
       })
     })
 
-    this.app.get('/xterm/xterm.css', (req, res) => {
-      res.sendFile(path.resolve(__dirname, 'console', 'node_modules', 'xterm', 'dist', 'xterm.css'))
-    })
-    
-    this.app.get('/xterm/xterm.js', (req, res) => {
-      res.sendFile(path.resolve(__dirname, 'console', 'node_modules', 'xterm', 'dist', 'xterm.js'))
-    })
-    
     this.app.use(Express.static(path.join(__dirname, 'console', 'dist')))
     
     this.app.get('*', function (request, response) {
