@@ -42,7 +42,7 @@ class DynamodbContainer extends ResourceContainer {
       const setupContainer = new Container(this.devServer, 'DynamoSetup', 'job')
       setupContainer.outputLogs = false
       setupContainer.daemon = false
-      setupContainer.getImage = () => { return 'mesosphere/aws-cli' }
+      setupContainer.getImage = () => { return 'mesosphere/aws-cli:latest' }
       setupContainer.getEnv = () => {
         return {
           AWS_ACCESS_KEY_ID: 'ABCDEF',
