@@ -7,10 +7,10 @@ class MysqlContainer extends ResourceContainer {
   }
 
   getImage () {
-    if (this.resource.params.version) {
-      return 'mysql:latest' // + this.resource.params.version
+    if (this.resource.settings.version) {
+      return `mysql:${this.resource.settings.version}`
     } else {
-      return 'mysql'
+      return 'mysql:latest'
     }
   }
 
