@@ -7,7 +7,11 @@ class PostgresqlContainer extends ResourceContainer {
   }
 
   getImage () {
-    return 'postgres:latest'
+    if (this.resource.settings.version) {
+      return `postgres:${this.resource.settings.version}-alpine`
+    } else {
+      return 'postgres:alpine'
+    }
   }
 
   getEnv () {
